refactor(layout-editor): separate view data and element registry
