+++ src/internal/ui/layoutEditor/preview/confirmDialog.js
@@ -1,0 +1,7 @@
+export const createConfirmDialogPreviewData = () => {
+  return {
+    resourceId: "",
+    confirmActions: {},
+    cancelActions: {},
+  };
+};
