+++ tests/layoutEditor/layoutEditorPersistence.test.js
@@ -7,14 +7,36 @@ import {
 
 describe("layoutEditorPersistence", () => {
   it("marks interaction fields as immediate-save changes", () => {
-    expect(shouldPersistLayoutEditorFieldImmediately("click")).toBe(true);
     expect(
-      shouldPersistLayoutEditorFieldImmediately("rightClick.payload.actions"),
+      shouldPersistLayoutEditorFieldImmediately({
+        name: "click",
+        itemType: "container",
+      }),
     ).toBe(true);
     expect(
-      shouldPersistLayoutEditorFieldImmediately("conditionalTextStyles"),
+      shouldPersistLayoutEditorFieldImmediately({
+        name: "rightClick.payload.actions",
+        itemType: "container",
+      }),
     ).toBe(true);
-    expect(shouldPersistLayoutEditorFieldImmediately("x")).toBe(false);
+    expect(
+      shouldPersistLayoutEditorFieldImmediately({
+        name: "conditionalTextStyles",
+        itemType: "text",
+      }),
+    ).toBe(true);
+    expect(
+      shouldPersistLayoutEditorFieldImmediately({
+        name: "paginationMode",
+        itemType: "container-ref-save-load-slot",
+      }),
+    ).toBe(true);
+    expect(
+      shouldPersistLayoutEditorFieldImmediately({
+        name: "x",
+        itemType: "container",
+      }),
+    ).toBe(false);
   });
 
   it("creates a patch payload for additive nested changes", () => {
